+++ tests/unit/helpers/typed-help-test.js
@@ -1,9 +1,9 @@
 import { typedHelp } from 'dummy/helpers/typed-help';
 import { module, test } from 'qunit';
 
-module('Unit | Helper | typed help', function() {
+module('Unit | Helper | typed help', function () {
   // Replace this with your real tests.
-  test('it works', function(assert) {
+  test('it works', function (assert) {
     let result = typedHelp([42]);
     assert.ok(result);
   });