+++ CHANGELOG.md
@@ -6,6 +6,22 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## [Unreleased]
 
+## [4.0.0-alpha.1] - 2020-01-18
+
+### Breaking 💥
+
+- Drop support for Node 8 ([#1017])
+- Don't configure Babel for TS transpilation ([#1018])
+
+### Under the hood 🚗
+
+- Upgrade [✨ Prettier ✨](https://prettier.io) to a version that supports optional chaining and nullish coalescing ([#1018])
+- Add test for optional chaining and nullish coalescing ([#1018])
+- Add test for class field declaration ([#1018])
+
+[#1017]: https://github.com/typed-ember/ember-cli-typescript/pull/1017
+[#1018]: https://github.com/typed-ember/ember-cli-typescript/pull/1018
+
 ## [3.1.4] - 2020-05-29
 
 ### Fixed 🔧
@@ -18,7 +34,6 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 [#1153]: https://github.com/typed-ember/ember-cli-typescript/pull/1153
 [3.1.4-deps-bumps]: https://github.com/typed-ember/ember-cli-typescript/pulls?q=is%3Apr+is%3Amerged+base%3Amaster+merged%3A2020-01-22T13%3A00%3A00-0600..2020-05-29T11%3A00%3A00-0500+chore%28deps%29+in%3Atitle+
-
 
 ## [3.1.3] - 2020-01-22
 
@@ -636,7 +651,12 @@ We now use Babel 7's support for TypeScript to build apps and addons. Most of th
 * Basic, semi-working functionality.
 
 [ember-cli-typify]: https://github.com/winding-lines/ember-cli-typify
-[unreleased]: https://github.com/typed-ember/ember-cli-typescript/compare/v3.1.3...HEAD
+[unreleased]: https://github.com/typed-ember/ember-cli-typescript/compare/v4.0.0-alpha.1...HEAD
+<!--
+  This is correctly compared against v3.1.3 because it was released on a branch
+  before v3.1.4 was released.
+-->
+[4.0.0-alpha.1]: https://github.com/typed-ember/ember-cli-typescript/compare/v3.1.3...v4.0.0-alpha.1
 [3.1.4]: https://github.com/typed-ember/ember-cli-typescript/compare/v3.1.3...v3.1.4
 [3.1.3]: https://github.com/typed-ember/ember-cli-typescript/compare/v3.1.2...v3.1.3
 [3.1.2]: https://github.com/typed-ember/ember-cli-typescript/compare/v3.1.1...v3.1.2